[DNM] - Fix UI lagging sometimes
I have no idea why this fixes the issue. In stock nougat, there is 2.62GB available to be used, however in lineage there appears to be 2.83GB. This might just be a GB -> GiB sort of thing, but this is still required to get smooth UI performance for now.

This is a very dirty hack to get to that level until I figure out why this causes such a drastic performance impact