Import sepolicy from lineage and slsi sources
Declare war on safetynet and play integrity attestation checks. The last thing I need to do for me to be "done" with 7420 is write working SELinux policies (and then go enforcing). It is tedious, boring, and quite frankly i'm not sure why I am doing this.

But I hope my efforts in this area can be used for other 7420 projects, if any are still alive in 2024 and beyond.